Disassociate products before deleting Manufacturer

+++ src/main/java/com/example/services/ManufacturerServiceImpl.java
@@ -1,9 +1,11 @@
 package com.example.services;
 
 import com.example.entities.Manufacturer;
+import com.example.entities.Product;
 import com.example.exception.EntityDeleteException;
 import com.example.exception.EntitySavingException;
 import com.example.repositories.ManufacturerRepository;
+import com.example.repositories.ProductRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,6 +21,7 @@ import java.util.Optional;
 @Service
 public class ManufacturerServiceImpl implements ManufacturerService {
     private ManufacturerRepository manufacturerRepo;
+    private ProductRepository productRepo;
 
     @Override
     public List<Manufacturer> findAll() {
@@ -115,7 +118,15 @@ public class ManufacturerServiceImpl implements ManufacturerService {
     @Override
     public void deleteById(Long id) {
         log.info("deleteById {}", id);
+        if (id == null || id <= 0)
+            throw new IllegalArgumentException("ID inválido");
+
+        Optional<Manufacturer> manufacturerOpt = manufacturerRepo.findById(id);
+        if (manufacturerOpt.isEmpty())
+            return;
+
         try {
+            disassociateProducts(manufacturerOpt.get());
             manufacturerRepo.deleteById(id);
             return;
         } catch (Exception e) {
@@ -123,5 +134,12 @@ public class ManufacturerServiceImpl implements ManufacturerService {
         }
 
         throw new EntityDeleteException("Error al borrar manufacturer.");
+    }
+
+    private void disassociateProducts(Manufacturer manufacturer) {
+        List<Product> products = productRepo.findAllByManufacturer_Cif(manufacturer.getCif());
+        for (Product product : products) {
+            product.setManufacturer(null);
+        }
     }
 }
